Split demo mode flag from demo sample data
The page layout checks isDemoMode() on every render, but importing it from
demo.ts dragged the band-name word pools and every demo sample array into
every page's module graph. demo.ts becomes a folder: mode.ts (the flag and
banner the layout needs), samples.ts (sample pools and the seeded listing
name generators), and overrides.ts (the per-form field maps and override
helpers). Also deletes DEMO_FIRST_NAMES and DEMO_SURNAMES, which had no
consumers.
