+++ src/features/admin/listings-edit.ts
@@ -41,7 +41,10 @@ import {
   updateListingAggregateValues,
 } from "#shared/db/listings.ts";
 import { settings } from "#shared/db/settings.ts";
-import { applyDemoOverrides, LISTING_DEMO_FIELDS } from "#shared/demo.ts";
+import {
+  applyDemoOverrides,
+  LISTING_DEMO_FIELDS,
+} from "#shared/demo/overrides.ts";
 import type { FormParams } from "#shared/form-data.ts";
 import {
   dimensionsOf,
