+++ src/features/admin/settings-stripe.ts
@@ -9,7 +9,7 @@ import {
   getWebhookUrl,
 } from "#routes/admin/settings-helpers.ts";
 import { settings } from "#shared/db/settings.ts";
-import { isDemoMode } from "#shared/demo.ts";
+import { isDemoMode } from "#shared/demo/mode.ts";
 import {
   detectStripeKeyMode,
   setupWebhookEndpoint,
