+++ src/shared/demo/samples.ts
@@ -1,10 +1,12 @@
 /**
- * Demo mode - replaces user-entered text with sample data to prevent PII storage.
- * Enable by setting DEMO_MODE=true environment variable.
+ * Sample data pools for demo mode and seeding.
+ *
+ * These arrays (and the generated listing pools) replace user-entered text
+ * when demo mode is on, and feed the database seeder. Only code that writes
+ * demo values should import this module — the per-render demo check lives in
+ * `#shared/demo/mode.ts` so pages never load these pools.
  */
 
-import { lazyRef } from "#fp";
-import { t } from "#i18n";
 import {
   DEFAULT_BAND_SEED,
   DEFAULT_DESCRIPTION_SEED,
@@ -13,35 +15,10 @@ import {
   generateDescriptions,
   generateVenueNames,
 } from "#shared/band-name-generator.ts";
-import { getEnv } from "#shared/env.ts";
-import type { FormParams } from "#shared/form-data.ts";
-import type { FieldValues } from "#shared/forms.tsx";
-import type { NamedResource } from "#shared/rest/resource.ts";
 
-// ---------------------------------------------------------------------------
-// Demo mode flag
-// ---------------------------------------------------------------------------
-
-const [getDemoMode, setDemoMode] = lazyRef(
-  () => getEnv("DEMO_MODE") === "true",
-);
-
-/** Check if demo mode is enabled */
-export const isDemoMode = (): boolean => getDemoMode();
-
-/** Reset cached demo mode value (for testing and cache invalidation) */
-export const resetDemoMode = (): void => setDemoMode(null);
-
-/**
- * Explicitly set demo mode on or off (for testing).
- * Bypasses Deno.env to avoid races between parallel test workers.
- */
-export const setDemoModeForTest = (enabled: boolean): void =>
-  setDemoMode(enabled);
-
-// ---------------------------------------------------------------------------
-// Sample data arrays
-// ---------------------------------------------------------------------------
+/** Pick a random element from an array */
+export const randomChoice = <T>(arr: readonly T[]): T =>
+  arr[Math.floor(Math.random() * arr.length)]!;
 
 /** Demo attendee names (full names for demo mode overrides) */
 export const DEMO_NAMES = [
@@ -149,197 +126,6 @@ export const DEMO_SERVICING_NAMES = [
   "Asbestos Survey",
   "Insurance Inspection",
   "Stage Lighting Focus",
-] as const;
-
-/** Demo first names for seed data (combined with surnames for more variety) */
-export const DEMO_FIRST_NAMES = [
-  "Alice",
-  "Bob",
-  "Carol",
-  "Dave",
-  "Eve",
-  "Frank",
-  "Grace",
-  "Henry",
-  "Ivy",
-  "Jack",
-  "Keiko",
-  "Liam",
-  "Maria",
-  "Nadia",
-  "Omar",
-  "Priya",
-  "Quincy",
-  "Rosa",
-  "Sven",
-  "Tomoko",
-  "Uche",
-  "Valentina",
-  "Wei",
-  "Xena",
-  "Yusuf",
-  "Zara",
-  "Aisling",
-  "Bjorn",
-  "Chiara",
-  "Dmitri",
-  "Elena",
-  "Fatima",
-  "George",
-  "Hana",
-  "Ibrahim",
-  "Jasmine",
-  "Kofi",
-  "Leila",
-  "Magnus",
-  "Nkechi",
-  "Olga",
-  "Pablo",
-  "Ravi",
-  "Sakura",
-  "Tariq",
-  "Uma",
-  "Viktor",
-  "Wanda",
-  "Yuki",
-  "Amara",
-  "Aiden",
-  "Bianca",
-  "Callum",
-  "Daphne",
-  "Emeka",
-  "Fiona",
-  "Gustavo",
-  "Haruki",
-  "Ingrid",
-  "Javier",
-  "Keira",
-  "Lorenzo",
-  "Mina",
-  "Niall",
-  "Oona",
-  "Petra",
-  "Rafael",
-  "Sienna",
-  "Theo",
-  "Ursula",
-  "Vijay",
-  "Wendy",
-  "Xiomara",
-  "Yolanda",
-  "Zane",
-  "Aoife",
-  "Benoit",
-  "Celine",
-  "Declan",
-  "Esme",
-  "Felix",
-  "Gemma",
-  "Hugo",
-  "Isla",
-  "Jules",
-  "Khadija",
-  "Lucian",
-  "Maeve",
-  "Nico",
-  "Orla",
-  "Piotr",
-  "Quinn",
-  "Rosario",
-  "Saoirse",
-  "Tomas",
-] as const;
-
-/** Demo surnames for seed data (combined with first names for more variety) */
-export const DEMO_SURNAMES = [
-  "Johnson",
-  "Smith",
-  "Williams",
-  "Brown",
-  "Davis",
-  "Miller",
-  "Wilson",
-  "Moore",
-  "Taylor",
-  "Anderson",
-  "Tanaka",
-  "O'Brien",
-  "Garcia",
-  "Petrova",
-  "Hassan",
-  "Sharma",
-  "Adams",
-  "Hernandez",
-  "Lindqvist",
-  "Sato",
-  "Okafor",
-  "Rossi",
-  "Zhang",
-  "Papadopoulos",
-  "Demir",
-  "Mbeki",
-  "Murphy",
-  "Eriksson",
-  "Bianchi",
-  "Volkov",
-  "Vasquez",
-  "Al-Rashid",
-  "Kamau",
-  "Yoshida",
-  "Diallo",
-  "Patel",
-  "Mensah",
-  "Khoury",
-  "Andersen",
-  "Eze",
-  "Novak",
-  "Reyes",
-  "Krishnan",
-  "Watanabe",
-  "Benmoussa",
-  "Devi",
-  "Szabo",
-  "Kowalski",
-  "Nakamura",
-  "Osei",
-  "Chen",
-  "Fitzgerald",
-  "Johansson",
-  "Kim",
-  "Larsson",
-  "Moreau",
-  "Nguyen",
-  "O'Sullivan",
-  "Park",
-  "Russo",
-  "Singh",
-  "Torres",
-  "Virtanen",
-  "Weber",
-  "Yamamoto",
-  "Zimmerman",
-  "Bergstrom",
-  "Costa",
-  "Dubois",
-  "Fischer",
-  "Gonzalez",
-  "Horvat",
-  "Ivanova",
-  "Jensen",
-  "Kapoor",
-  "Li",
-  "Martinez",
-  "Nielsen",
-  "Oduya",
-  "Pereira",
-  "Ramirez",
-  "Svensson",
-  "Takahashi",
-  "Ueda",
-  "Varga",
-  "Walsh",
-  "Xu",
-  "Yilmaz",
 ] as const;
 
 /** Demo email addresses */
@@ -549,107 +335,3 @@ export const DEMO_TERMS = [
   "By registering you agree to our community listing guidelines. Refunds available up to 48 hours before the listing.",
   "All attendees must follow the venue code of conduct. Photography may take place at this listing.",
 ] as const;
-
-// ---------------------------------------------------------------------------
-// Field mapping type and pre-built mappings
-// ---------------------------------------------------------------------------
-
-/** Maps form field names to arrays of possible demo values */
-export type DemoFieldMap = Record<string, readonly string[]>;
-
-/** Attendee PII fields */
-export const ATTENDEE_DEMO_FIELDS: DemoFieldMap = {
-  address: DEMO_ADDRESSES,
-  email: DEMO_EMAILS,
-  name: DEMO_NAMES,
-  phone: DEMO_PHONES,
-  special_instructions: DEMO_SPECIAL_INSTRUCTIONS,
-};
-
-/** Servicing-event fields — name only, and a servicing reason rather than a
- * person's name, so demo mode doesn't turn "Boiler Service" into "Bob Smith". */
-export const SERVICING_DEMO_FIELDS: DemoFieldMap = {
-  name: DEMO_SERVICING_NAMES,
-};
-
-/** Listing metadata fields */
-export const LISTING_DEMO_FIELDS: DemoFieldMap = {
-  description: DEMO_LISTING_DESCRIPTIONS,
-  location: DEMO_LISTING_LOCATIONS,
-  name: DEMO_LISTING_NAMES,
-};
-
-/** Group name and description fields */
-export const GROUP_DEMO_FIELDS: DemoFieldMap = {
-  description: DEMO_GROUP_DESCRIPTIONS,
-  name: DEMO_GROUP_NAMES,
-};
-
-/** Holiday name field */
-export const HOLIDAY_DEMO_FIELDS: DemoFieldMap = {
-  name: DEMO_HOLIDAY_NAMES,
-};
-
-/** Site homepage fields */
-export const SITE_HOME_DEMO_FIELDS: DemoFieldMap = {
-  homepage_text: DEMO_PAGE_TEXT,
-  website_title: DEMO_WEBSITE_TITLES,
-};
-
-/** Attendee Logistics tab fields: the address is masked like the attendee
- * form's, and the pinned latitude/longitude are cleared outright — an exact
- * real-world location is PII even beside a masked address. */
-export const LOGISTICS_DEMO_FIELDS: DemoFieldMap = {
-  address: DEMO_ADDRESSES,
-  lat: [""],
-  lng: [""],
-};
-
-/** Site contact page fields */
-export const SITE_CONTACT_DEMO_FIELDS: DemoFieldMap = {
-  contact_page_text: DEMO_PAGE_TEXT,
-};
-
-/** Terms and conditions field */
-export const TERMS_DEMO_FIELDS: DemoFieldMap = {
-  terms_and_conditions: DEMO_TERMS,
-};
-
-// ---------------------------------------------------------------------------
-// Override logic
-// ---------------------------------------------------------------------------
-
-/** Pick a random element from an array */
-export const randomChoice = <T>(arr: readonly T[]): T =>
-  arr[Math.floor(Math.random() * arr.length)]!;
-
-/**
- * Replace form field values with demo data when demo mode is active.
- * Only replaces fields that are present and non-empty in the form.
- * Mutates and returns the same URLSearchParams for chaining.
- */
-export const applyDemoOverrides = (
-  form: FormParams,
-  mapping: DemoFieldMap,
-): FormParams => {
-  if (!isDemoMode()) return form;
-  for (const [field, values] of Object.entries(mapping)) {
-    if (form.has(field) && form.get(field) !== "") {
-      form.set(field, randomChoice(values));
-    }
-  }
-  return form;
-};
-
-/** Wrap a named resource so create/update apply demo overrides to the form */
-export const wrapResourceForDemo = <R, I, V extends FieldValues = FieldValues>(
-  resource: NamedResource<R, I, V>,
-  mapping: DemoFieldMap,
-): NamedResource<R, I, V> => ({
-  ...resource,
-  create: (form) => resource.create(applyDemoOverrides(form, mapping)),
-  update: (id, form) => resource.update(id, applyDemoOverrides(form, mapping)),
-});
-
-/** Demo mode banner HTML */
-export const DEMO_BANNER = `<div class="demo-banner">${t("guide.demo_mode_notice")}</div>`;
