+++ test/lib/server-demo-reset.test.ts
@@ -3,7 +3,7 @@ import { afterEach, beforeEach, describe, it as test } from "@std/testing/bdd";
 import { handleRequest } from "#routes";
 import { imagesTable } from "#shared/db/images.ts";
 import { listingsTable } from "#shared/db/listings.ts";
-import { setDemoModeForTest } from "#shared/demo.ts";
+import { setDemoModeForTest } from "#shared/demo/mode.ts";
 import { runWithStorageConfig } from "#shared/storage.ts";
 import { nonEmptyString } from "#shared/validation/string.ts";
 import {
