+++ test/lib/server-listings/edit-post-basics.test.ts
@@ -3,7 +3,7 @@ import { expect } from "@std/expect";
 import { afterEach, describe, it as test } from "@std/testing/bdd";
 import { handleRequest } from "#routes";
 import { getListingWithCount } from "#shared/db/listings.ts";
-import { setDemoModeForTest } from "#shared/demo.ts";
+import { setDemoModeForTest } from "#shared/demo/mode.ts";
 import {
   adminFormPost,
   createTestListing,
